manager/requirements.txt
fire==0.1.3
psycopg2==2.7.7
pyyaml==5.1.1
bcrypt==3.1.7